feat: port over polyfill updates from https://github.com/boltdesignsystem/bolt/pull/1579
