+++ packages/base-polyfills/index.js
@@ -14,6 +14,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
  */
 
 import 'regenerator-runtime/runtime';
+import '@webcomponents/custom-elements/src/native-shim.js';
 import 'core-js/modules/es.array.find';
 import '@bolt/polyfills/platform/remove-polyfill.js';
 import '@bolt/polyfills/platform/es6-misc';
@@ -26,3 +27,43 @@ import '@webcomponents/custom-elements/src/custom-elements.js';
 import '@webcomponents/url/url.js';
 import '@bolt/polyfills/platform/baseuri';
 import '@bolt/polyfills/platform/unresolved';
+import 'element-closest';
+import 'whatwg-fetch';
+import 'mdn-polyfills/Node.prototype.prepend';
+import 'mdn-polyfills/Node.prototype.replaceWith'; // used in dropdown
+import 'core-js/modules/es.array.iterator';
+import 'core-js/modules/es.array.from';
+import 'core-js/modules/es.string.starts-with';
+import 'core-js/modules/es.array.includes';
+import 'core-js/modules/es.array.for-each';
+import 'core-js/modules/es.object.assign';
+import 'core-js/modules/es.string.includes';
+import 'core-js/modules/es.string.repeat';
+import WeakSet from '@ungap/weakset';
+window.WeakSet = WeakSet;
+
+import smoothscroll from 'smoothscroll-polyfill';
+
+// kick off the polyfill!
+smoothscroll.polyfill();
+
+/**
+ * closest() polyfill
+ * @link https://developer.mozilla.org/en-US/docs/Web/API/Element/closest#Polyfill
+ */
+if (window.Element && !Element.prototype.closest) {
+  Element.prototype.closest = function(s) {
+    var matches = (this.document || this.ownerDocument).querySelectorAll(s),
+      i,
+      el = this;
+    do {
+      i = matches.length;
+      while (--i >= 0 && matches.item(i) !== el) {}
+    } while (i < 0 && (el = el.parentElement));
+    return el;
+  };
+}
+
+if (window.NodeList && !NodeList.prototype.forEach) {
+  NodeList.prototype.forEach = Array.prototype.forEach;
+}
