+++ packages/core/polyfills/index.js
@@ -1,48 +1,9 @@
+// @todo: remove this import once the ES Module work is merged in + migration place for existing external use of polyfills in place
 import '@bolt/polyfills';
-import 'element-closest';
-import 'whatwg-fetch';
-import 'mdn-polyfills/Node.prototype.prepend';
-import 'mdn-polyfills/Node.prototype.replaceWith'; // used in dropdown
-import 'core-js/modules/es.array.iterator';
-import 'core-js/modules/es.array.from';
-import 'core-js/modules/es.string.starts-with';
-import 'core-js/modules/es.array.includes';
-import 'core-js/modules/es.array.for-each';
-import 'core-js/modules/es.object.assign';
-import 'core-js/modules/es.string.includes';
-import 'core-js/modules/es.string.repeat';
-import WeakSet from '@ungap/weakset';
-window.WeakSet = WeakSet;
-
-import smoothscroll from 'smoothscroll-polyfill';
-
-// kick off the polyfill!
-smoothscroll.polyfill();
-
-/**
- * closest() polyfill
- * @link https://developer.mozilla.org/en-US/docs/Web/API/Element/closest#Polyfill
- */
-if (window.Element && !Element.prototype.closest) {
-  Element.prototype.closest = function(s) {
-    var matches = (this.document || this.ownerDocument).querySelectorAll(s),
-      i,
-      el = this;
-    do {
-      i = matches.length;
-      while (--i >= 0 && matches.item(i) !== el) {}
-    } while (i < 0 && (el = el.parentElement));
-    return el;
-  };
-}
-
-if (window.NodeList && !NodeList.prototype.forEach) {
-  NodeList.prototype.forEach = Array.prototype.forEach;
-}
 
 let polyfills = [];
 
-// // Detect Shadow Dom Support
+// Detect Shadow Dom Support
 if (
   !(
     'attachShadow' in Element.prototype && 'getRootNode' in Element.prototype
@@ -82,8 +43,6 @@ export const polyfillLoader = new Promise(resolve => {
   ) {
     resolve();
   } else {
-    import('@webcomponents/custom-elements/src/native-shim.js').then(() => {
-      resolve();
-    });
+    resolve();
   }
 });
